+++ CHANGES.md
@@ -7,6 +7,7 @@ Changes in 4.9
 - Accessibility: add missing input label in AssignOtherInfo.php
 - PROGRAM_CONFIG table: Add Allow Teachers to edit gradebook grades for past quarters option in Update.fnc.php, rosariosis.sql & Configuration.php, sponsored by Santa Cecilia school
 - Accessibility: add missing input label in Widgets.fnc.php
+- Do not search custom Fields of the Files type in Search.fnc.php & Preferences.php
 
 Changes in 4.8
 --------------