+++ CHANGES.md
@@ -8,6 +8,7 @@ Changes in 4.9
 - PROGRAM_CONFIG table: Add Allow Teachers to edit gradebook grades for past quarters option in Update.fnc.php, rosariosis.sql & Configuration.php, sponsored by Santa Cecilia school
 - Accessibility: add missing input label in Widgets.fnc.php
 - Do not search custom Fields of the Files type in Search.fnc.php & Preferences.php
+- Move Course Period option inputs + Takes Attendance input to functions in Courses.php & Courses.fnc.php
 
 Changes in 4.8
 --------------